+++ subtest/src/lib.rs
@@ -410,19 +410,27 @@
 //! * If it is a nested function with added attributes like `#[should_panic]`: **Specify a `#[test]` attribute as well!** (see [Omit or Override Attributes, Parameters, Return Types](#omit-or-override-attributes-parameters-return-types))
 //! * Always put other attributes **after** `#[subtest]`
 //!
-//! In case you do ever forget the `#[test]` attribute, you will get the same compiler warning as if you forgot to annotate a regular test function with `#[test]`:
+//! In case you do ever forget the `#[test]` attribute, you will get a compiler error:
 //!
 //! ```text
-//! warning: function `parent` is never used
-//!  --> subtest/tests/expand/missing_test_attr.rs:2:4
+//! error: function is missing a test attribute, such as #[test], #[tokio::test] or #[rstest] - add one below #[subtest]
+//!  --> tests/ui/fail/missing_test_attr.rs:4:4
 //!   |
-//! 2 | fn parent() {
+//! 4 | fn parent() {
 //!   |    ^^^^^^
-//!   |
-//!   = note: `#[warn(dead_code)]` (part of `#[warn(unused)]`) on by default
 //! ```
 //!
-//! I recommend automatically running clippy with deny-warnings turned on (`cargo clippy --locked --all-targets --all-features -- -D warnings`), which catches these issues fairly quickly.
+//! `subtest` determines test attributes by checking whether they end with `test`.
+//! If your test attribute of choice does not, you can opt out of the check via:
+//!
+//! ```
+//! # use subtest::subtest;
+//! #[subtest(allow_missing_test_attribute)]
+//! // your weirdly-named test attribute goes here!
+//! fn top_level() {
+//!     // ...
+//! }
+//! ```
 //!
 //! ## Ambiguous macro import
 //!
@@ -607,10 +615,10 @@ use subtest_impl::expand_subtest_main_fn;
 ///
 /// 1. On the top-level test function which contains the subtests
 ///    * after `#[subtest]`, you **have to** add a test attribute of the testing framework you intend to use, such as regular `#[test]`, `#[tokio::test]`, `#[rstest]`, etc.
+///    * optional arguments:
+///       * `#[subtest(allow_missing_test_attribute)]` - disables the compiler check for whether a test attribute is present
 /// 2. On each individual subtest function
 ///    * here, you can omit test attributes, parameters, and return types - they are inherited from the parent test function
-///
-/// `#[subtest]` takes no arguments.
 ///
 /// For more information, refer to the [crate-level documentation](crate).
 ///