+++ subtest/tests/expand/missing_test_attr.rs
@@ -1,4 +1,4 @@
-#[subtest::subtest]
+#[subtest::subtest(allow_missing_test_attribute)]
 fn parent() {
     #[subtest]
     fn child() {}